client channel: create a header file for internal interfaces
