+++ BUILD
@@ -2763,6 +2763,7 @@ grpc_cc_library(
         "//src/core:ext/filters/client_channel/client_channel.h",
         "//src/core:ext/filters/client_channel/client_channel_channelz.h",
         "//src/core:ext/filters/client_channel/client_channel_factory.h",
+        "//src/core:ext/filters/client_channel/client_channel_internal.h",
         "//src/core:ext/filters/client_channel/client_channel_service_config.h",
         "//src/core:ext/filters/client_channel/config_selector.h",
         "//src/core:ext/filters/client_channel/connector.h",
@@ -2770,7 +2771,6 @@ grpc_cc_library(
         "//src/core:ext/filters/client_channel/global_subchannel_pool.h",
         "//src/core:ext/filters/client_channel/health/health_check_client.h",
         "//src/core:ext/filters/client_channel/http_proxy.h",
-        "//src/core:ext/filters/client_channel/lb_call_state_internal.h",
         "//src/core:ext/filters/client_channel/lb_policy/child_policy_handler.h",
         "//src/core:ext/filters/client_channel/lb_policy/oob_backend_metric.h",
         "//src/core:ext/filters/client_channel/lb_policy/oob_backend_metric_internal.h",
