+++ tools/doxygen/Doxyfile.core.internal
@@ -897,6 +897,7 @@ src/core/ext/filters/client_channel/client_channel_channelz.cc \
 src/core/ext/filters/client_channel/client_channel_channelz.h \
 src/core/ext/filters/client_channel/client_channel_factory.cc \
 src/core/ext/filters/client_channel/client_channel_factory.h \
+src/core/ext/filters/client_channel/client_channel_internal.h \
 src/core/ext/filters/client_channel/client_channel_plugin.cc \
 src/core/ext/filters/client_channel/client_channel_service_config.cc \
 src/core/ext/filters/client_channel/client_channel_service_config.h \
@@ -911,7 +912,6 @@ src/core/ext/filters/client_channel/health/health_check_client.cc \
 src/core/ext/filters/client_channel/health/health_check_client.h \
 src/core/ext/filters/client_channel/http_proxy.cc \
 src/core/ext/filters/client_channel/http_proxy.h \
-src/core/ext/filters/client_channel/lb_call_state_internal.h \
 src/core/ext/filters/client_channel/lb_policy/address_filtering.cc \
 src/core/ext/filters/client_channel/lb_policy/address_filtering.h \
 src/core/ext/filters/client_channel/lb_policy/backend_metric_data.h \
